miniC: Improve parser for function call
If there is a function call like `hello(x)`, it causes ambiguous whether it is function call or declaration.
To make the decision, we have to keep track of types and declarators which are declared in environment when parsing. Reaching `hello`, we look up in the environment, we can identify `hello`